+++ ddm/logging/api/serializers.py
@@ -4,27 +4,25 @@ from ddm.logging.models import EventLogEntry, ExceptionLogEntry
 
 
 class EventLogEntrySerializer(serializers.ModelSerializer):
-    date = serializers.DateTimeField(format='%Y-%m-%d, %H:%M')
+    date = serializers.DateTimeField(format="%Y-%m-%d, %H:%M")
 
     class Meta:
         model = EventLogEntry
-        fields = ['date', 'description', 'message']
+        fields = ["date", "description", "message"]
 
 
 class ExceptionLogEntrySerializer(serializers.ModelSerializer):
-    date = serializers.DateTimeField(format='%Y-%m-%d, %H:%M')
-    participant = serializers.CharField(
-        source='participant.external_id', default=None
-    )
-    blueprint = serializers.CharField(source='blueprint.name', default=None)
+    date = serializers.DateTimeField(format="%Y-%m-%d, %H:%M")
+    participant = serializers.CharField(source="participant.external_id", default=None)
+    blueprint = serializers.CharField(source="blueprint.name", default=None)
 
     class Meta:
         model = ExceptionLogEntry
         fields = [
-            'date',
-            'participant',
-            'exception_type',
-            'raised_by',
-            'blueprint',
-            'message',
+            "date",
+            "participant",
+            "exception_type",
+            "raised_by",
+            "blueprint",
+            "message",
         ]