+++ ddm/logging/apps.py
@@ -3,7 +3,7 @@ from django.utils.translation import gettext_lazy as _
 
 
 class DDMLoggingConfig(AppConfig):
-    default_auto_field = 'django.db.models.AutoField'
-    name = 'ddm.logging'
-    label = 'ddm_logging'
-    verbose_name = _('DDM Logging')
+    default_auto_field = "django.db.models.AutoField"
+    name = "ddm.logging"
+    label = "ddm_logging"
+    verbose_name = _("DDM Logging")